+++ CHANGELOG.md
@@ -10,6 +10,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 -   Benchmark names can now include URL query strings (e.g. `tach
     my-benchmark?foo=1&baz=2`) which will be included as-is in the launched URL.
 
+-   Variants no longer exist. Use URL query strings intead (see above).
+
 -   `--manual` mode no longer shows benchmark data (but it should only be used
     for testing the web server anyway since it has no statistical significance).